summariser/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
PyPDF2
python-docx
python-dotenv
google-generativeai